pyproject.toml
[tool.poetry]
name = "shamrock-k8s"
version = "0.0.0"
description = "Shamrock Kubernetes Cluster Manifests"
authors = ["Brian Lu <me@greencappuccino.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
taskipy = "^1.10.3"
pyyaml = "^6.0"

[tool.taskipy.tasks]
pre-commit-hook = "pre-commit install"
pre-commit = "pre-commit run --all-files"
inflate = "python inflate.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
